refactor pack.c to also compile on target
